Renamed behaviour to behavior.
