+++ test/ReentrancyGuard.test.js
@@ -16,7 +16,7 @@ contract('ReentrancyGuard', function () {
     await expectThrow(reentrancyMock.countAndCall(attacker.address));
   });
 
-  // The following are more side-effects than intended behaviour:
+  // The following are more side-effects than intended behavior:
   // I put them here as documentation, and to monitor any changes
   // in the side-effects.
