+++ test/token/ERC20/BurnableToken.behavior.js
