+++ test/token/ERC20/CappedToken.behavior.js
