+++ test/token/ERC20/MintableToken.behavior.js
