+++ test/token/ERC721/ERC721BasicToken.behavior.js
