+++ test/token/ERC721/ERC721BasicToken.test.js
@@ -1,5 +1,5 @@
-const { shouldBehaveLikeERC721BasicToken } = require('./ERC721BasicToken.behaviour');
-const { shouldBehaveLikeMintAndBurnERC721Token } = require('./ERC721MintBurn.behaviour');
+const { shouldBehaveLikeERC721BasicToken } = require('./ERC721BasicToken.behavior');
+const { shouldBehaveLikeMintAndBurnERC721Token } = require('./ERC721MintBurn.behavior');
 
 const BigNumber = web3.BigNumber;
 const ERC721BasicToken = artifacts.require('ERC721BasicTokenMock.sol');
