+++ test/token/ERC721/ERC721MintBurn.behavior.js
